chore: applying DRY principle
