+++ app/action/personal.js
@@ -1,18 +1,6 @@
-const config = require('../../config');
 const ConnectDB = require('../util/connectDB');
 const getResponse = require('../util/getResponse');
-
-
-const upsertColumns = (cols) => 
-	config.dbType == 'sqlite'
-	? cols.map(c => `${c}=excluded.${c}`).join(',')
-	: cols.map(c => `${c}=VALUES(${c})`).join(',')
-
-const upsertClause = (cols) => (
-	config.dbType == 'sqlite'
-		? 'ON CONFLICT (gid) DO UPDATE SET '
-		: 'ON DUPLICATE KEY UPDATE '
-	) + upsertColumns(cols)
+const { upsertClause } = require('../util/queryCommon');
 
 /**
  * 
@@ -30,7 +18,7 @@ const personal = async (req, res) => {
 		const result = await conn.query(
 			`INSERT INTO gallery_personal (${cols.join(',')})
 				VALUES (${paramStr})
-				${upsertClause(cols)}`,
+				${upsertClause(cols, 'gid')}`,
 			cols.map(c => req.body[c])
 		);
 		return res.json(getResponse(result, 200, 'success', {}));
