+++ app/action/search.js
@@ -2,11 +2,9 @@ const ConnectDB = require('../util/connectDB');
 const getResponse = require('../util/getResponse');
 const matchExec = require('../util/matchExec');
 const { categoryMap } = require('../util/category');
-const queryPersonal = require('../util/queryPersonal');
-const queryTags = require('../util/queryTags');
-const queryTorrents = require('../util/queryTorrents');
 const normalizedTag = require('../util/normalizedTag');
 const config = require('../../config');
+const { populateGalleryData } = require('../util/queryCommon');
 
 const search = async (req, res) => {
 	let {
@@ -258,27 +256,12 @@ const search = async (req, res) => {
 			return res.json(getResponse([], 200, 'success', { total }));
 		}
 
-		await populateAdditional(conn, result);
+		await populateGalleryData(conn, result);
 
 		return res.json(getResponse(result, 200, 'success', { total }));
 	} finally {
 		conn.end();
 	}
-};
-
-// TODO: move into reusable util to share with other app/actions
-const populateAdditional = async (conn, result) => {
-	const gids = result.map(e => e.gid);
-	const rootGids = result.map(e => e.root_gid).filter(e => e);
-	const gidTags = await queryTags(conn, gids);
-	const gidTorrents = await queryTorrents(conn, rootGids);
-	const personal = config.features.personal && await queryPersonal(conn, gids);
-
-	result.forEach(e => {
-		e.tags = gidTags[e.gid] || [];
-		e.torrents = gidTorrents[e.root_gid] || [];
-		e.personal = personal[e.gid] || {};
-	});
 };
 
 module.exports = search;
