+++ app/action/searchImage.js
@@ -1,9 +1,8 @@
 const phash = require("sharp-phash");
 const ConnectDB = require('../util/connectDB');
 const getResponse = require('../util/getResponse');
-const queryTags = require("../util/queryTags");
-const queryTorrents = require("../util/queryTorrents");
 const config = require("../../config");
+const { populateGalleryData } = require("../util/queryCommon");
 
 const searchImage = async (req, res) => {
 	if (!req.file) {
@@ -40,15 +39,7 @@ const searchImage = async (req, res) => {
 		);
 
 		if (out !== 'gid') {
-			const gids = result.map(e => e.gid);
-			const rootGids = result.map(e => e.root_gid).filter(e => e);
-			const gidTags = await queryTags(conn, gids);
-			const gidTorrents = await queryTorrents(conn, rootGids);
-
-			result.forEach(e => {
-				e.tags = gidTags[e.gid] || [];
-				e.torrents = gidTorrents[e.root_gid] || [];
-			});
+			await populateGalleryData(conn, result);
 		}
 
 		return res.json(getResponse(result, 200, 'success', { total: result.length }));
